Backend/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
bidict==0.23.1
billiard==4.2.0
bitsandbytes==0.42.0
blinker==1.8.2
breadability==0.1.20
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
confection==0.1.4
ConfigArgParse==1.7
confusable-homoglyphs==3.3.1
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.5.14
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
Django==4.2.15
django-allauth==64.1.0
django-cors-headers==4.4.0
django-email-verification==0.3.3
django-extensions==3.2.3
django-mysql==4.14.0
django-registration==3.4
django-registration-redux==2.13
Django-Verify-Email==2.0.3
djangorestframework==3.15.2
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.2
exceptiongroup==1.2.1
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
futures==3.0.5
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
geventhttpclient==2.0.12
goslate==1.5.4
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h2==3.2.0
h5py==3.8.0 
hpack==3.0.0
hstspreload==2024.6.1
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.13.1
kiwisolver==1.4.7
kombu==5.4.0
langcodes==3.4.0
langsmith==0.1.81
language_data==1.2.0
Lazify==0.4.0
libclang==18.1.1
literalai==0.0.604
locust==2.25.0
lxml==5.2.2
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mathparse==0.1.2
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-semantic-conventions==0.46b0
opt-einsum==3.3.0
orjson==3.10.5
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
peewee==3.17.5
pillow==10.4.0
Pint==0.21.1
preshed==3.0.9
protobuf==4.25.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
PyDictionary==2.0.1
PyJWT==2.8.0
pymongo==3.13.0
PyMySQL==1.1.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==5.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
roundrobin==0.0.4
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.0.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.2.19
sqlparse==0.5.1
srsly==2.4.8
starlette==0.37.2
sumy==0.11.0
sympy==1.12.1
syncer==2.0.3
tenacity==8.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
transformers==4.41.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.5.0
tzdata==2024.1
ujson==5.10.0
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
uvloop==0.19.0
validators==0.33.0
vine==5.1.0
wasabi==1.1.2
watchfiles==0.20.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.40
zipp==3.19.1
zope.event==5.0
zope.interface==7.0.1
